notebooks/rowan/requirements.txt
google-cloud-bigquery==3.12.0
google-cloud-bigquery-storage==2.22.0
google-cloud-storage==2.12.0
Jinja2==3.1.2
pandas==2.1.1
pyarrow==13.0.0
PyYAML==6.0.1
